feat(ci): CI-running tests
This implements some lo2s tests that can and will actually run in the
CI.

While most of the perf hardware event based tests can not run, the
tracepoint based tests run, which should give us a pretty wide covering
of basic lo2s functionality at least.

In this step, I replace the fixtures with checks that make ctest skip
the test if the pre-requisites don't exist (For example if the event we
want to use does not exist per `perf`).

I acknowledge that we initially chosen fixtures for these pre-requisites
on purpose and this encodes them a lot more implicitly, however the
SKIP_RETURN_CODE feature is the only way to tell ctest that a tests
failure (including a fixtures one) is benign.

Per default ctest will still loudly fail if a test fails for the reason
that a pre-requisite is not there, alerting the user with a printed
warning what causes the test failure.

Specifically for the CI, there is LO2S_TESTS_MAY_SKIP, which is
intentionally undocumented and which makes unexecutable tests skip.

+++ cmake/tests/block_io/block_io.sh
@@ -1,16 +1,27 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
+SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
+
+if ! bash $SCRIPT_DIR/../paranoid.sh -1; then
+	echo "block I/O recording test needs kernel.perf_event_paranoid=-1" >&2
+	exit 127
+fi
+
+if ! bash $SCRIPT_DIR/../test_sys_kernel_tracing.sh; then
+	echo "/sys/kernel/tracing needs to be read-accessible for block I/O test" >&2
+	exit 127
+fi
+
 rm -rf test_trace
 
-./lo2s -c 1 --block-io --output-trace test_trace -- bash -c "echo "TEST!" > test_file && sleep 1 && sync"
+./lo2s --no-instruction-sampling --block-io --output-trace test_trace -- bash -c "echo "TEST!" > test_file && sleep 1 && sync"
 
 rm test_file
 
-if otf2-print test_trace/traces.otf2 |  grep "IO_OPERATION" > /dev/null
-then
-    exit 0
+if otf2-print test_trace/traces.otf2 | grep "IO_OPERATION" >/dev/null; then
+	exit 0
 else
-    echo "Trace did not contain IO operations!"
-    exit 1
+	echo "Trace did not contain IO operations!"
+	exit 1
 fi
