+++ cmake/tests/process_sampling/process_sampling.sh
@@ -1,14 +1,25 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
+SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
+
+if ! bash $SCRIPT_DIR/../paranoid.sh 2; then
+	echo "sampling in process mode test needs kernel.perf_event_paranoid=2" >&2
+	exit 127
+fi
+
+if ! bash $SCRIPT_DIR/../has_req_perf_events.sh; then
+	echo "sampling in process mode test needs access to the 'instructions' perf event!" >&2
+	exit 127
+fi
+
 rm -rf test_trace
 
-./lo2s -c 1  --output-trace test_trace -- sleep 1
+./lo2s -c 1 --output-trace test_trace -- sleep 1
 
-if  otf2-print test_trace/traces.otf2 | grep  "SAMPLE" > /dev/null
-then
-    exit 0
+if otf2-print test_trace/traces.otf2 | grep "SAMPLE" >/dev/null; then
+	exit 0
 else
-    echo "Trace did not contain calling context samples in process mode!"
-    exit 1
+	echo "Trace did not contain calling context samples in process mode!"
+	exit 1
 fi
